requirements.txt
streamlit
genai
pandas
numpy
python-dotenv
google.generativeai
